task 4

+++ source/zad4/klient.c
@@ -10,23 +10,22 @@
 
 #define MAXBUF 80
 
-void child_func(int childnum, char *servIP, int port);
+void child_func (int childnum, char *servIP, int port);
 
-int main(int argc, char *argv[])
-{
+int main (int argc, char *argv[]) {
     int nchildren = 1;
-    char *servIP;
+    char * servIP;
     int port;
     int pid;
     int i;
 
     if (argc < 4) {
-      fprintf(stderr, "Uzycie: %s <serwer> <port> <liczba klientow>\n",
-              argv[0]);
-      exit(1);
+        fprintf(stderr, "Uzycie: %s <serwer> <port> <liczba klientow>\n", argv[0]);
+        exit(1);
     }
-    servIP=argv[1];
-    port=atoi(argv[2]);
+    
+    servIP = argv[1];
+    port = atoi(argv[2]);
     nchildren = atoi(argv[3]);
     
     for (i = 0; i < nchildren; i++) {
@@ -35,16 +34,16 @@ int main(int argc, char *argv[])
             exit(0);
         }
     }
+    
     wait(NULL);
+    
     return 0;
 }
 
-void child_func(int childnum, char *servIP, int port)
-{
+void child_func (int childnum, char *servIP, int port) {
     int sock;
     struct sockaddr_in sAddr;
     char buffer[MAXBUF];
-    
 
     sock = socket(PF_INET, SOCK_STREAM, 0);
     sAddr.sin_family = AF_INET;
@@ -52,17 +51,21 @@ void child_func(int childnum, char *servIP, int port)
     sAddr.sin_port = htons(port);
     
     if (connect(sock, (const struct sockaddr *) &sAddr, sizeof(sAddr)) != 0) {
-      perror("klient");
-      return;
+        perror("klient");
+        return;
     }
-      
+
     snprintf(buffer, MAXBUF, "Dane od klienta %i.", childnum);
+    
     sleep(1);
-    printf("Potomek %i wyslal %i znakow\n", childnum, 
-            write(sock, buffer, strlen(buffer)));
+    
+    printf("Potomek %i wyslal %i znakow\n", childnum, (int) write(sock, buffer, strlen(buffer)));
+    
     sleep(1);
-    printf("Potomek %i otrzymal %i znakow\n", childnum, 
-            read(sock, buffer, MAXBUF));
+    
+    printf("Potomek %i otrzymal %i znakow\n", childnum, (int) read(sock, buffer, MAXBUF));
+    
     sleep(1);
+    
     close(sock);
 }
